requirements.txt
docx==0.2.4
Faker==8.8.1
ffmpeg==1.4
ffmpeg_python==0.2.0
openpyxl==3.0.10
Pillow==9.3.0
python_docx==0.8.11
reportlab==3.6.12
xlwt==1.3.0
